Update igc_lib.py
_compute_takeoff_landing failed on this flight: https://www.onlinecontest.org/olc-3.0/gliding/flightinfo.html?dsId=7571663
Takeoff is on roughly on the 13th fix. However, going through compute_flight, the first fix is automatically going to be set to fix.flying = True, even if output = 0 for the first Fix. This is due to ignore_next_downtime and apply_next_downtime both being set to False. 
I have done no thorough testing, but setting apply_next_downtime to False for the start finds the correct takeoff_fix.